+++ PickleBall.Domain/Entities/Transaction.cs
@@ -8,7 +8,7 @@ namespace PickleBall.Domain.Entities
         public Guid UserId { get; set; }
         public Guid WalletId { get; set; }
         public string Status { get; set; }
-        public BigInteger Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Description { get; set; }
         public Guid BookingId { get; set; }
         public DateTimeOffset CreatedOnUtc { get; set; }
